0.2.0
- Added console commands: PlayerFactionLearnBlueprints, PlayerFactionUpdateVariants, PrintPlayerFactionFleets
- Fixed bug with clear roles method
- Added Intel Plugin for management
- Added settings
"requiredPriorityCheck": true,` - Setting to govern whether to check for priority tick in doctrine screen when updating variants. If
false, variants for player tech ships will be added even if not ticked
"requiredCommissionCheck": true` - Setting to govern whether to check for active commission as condition to generate blueprint

- Added max_percentage settings
The aggregated percentage that your ships will sum up to during fleet generation for each role. If role is missing, it will default to 100
"COMBAT_LARGE": 100,
"COMBAT_MEDIUM": 100